ok, j'ai arranger le bug precedent. Matlab etait en mode ou les donnees etaient en mode uint8 par defaut. Donc, l<iterateur sur le signal ne pouvait pas atteindre plus que 255. Mais le signal a 16 000+ valeurs!! Rendu a 256, le programme plantait, mais silencieusement. il retournait donc le resultat des 255 premiers element mais cetait pas complet!

+++ tp1/manipulation2.m
@@ -1,5 +1,5 @@
 signal = imread('lena.gif'); % on charge l'image
 signal = signal(:)'; % on transforme la matrice 2D vers un vecteur 1D.
 
-signal_compresse = lz77(signal, 32, 16);
+signal_compresse = lz77(uint8(signal), 32, 16);
 taux_compression = length(signal_compresse) / length(signal);
